начала задание с использованием CommonJS, вынесла загрузку отзывов в отдельный файл
